Fix #4: NoClassDefFoundError: kotlin/time/TimeSource$Monotonic

+++ src/main/kotlin/org/asarkar/gradle/Printer.kt
@@ -1,8 +1,8 @@
 package org.asarkar.gradle
 
 data class PrinterInput(
-        val buildDuration: Double,
-        val taskDurations: List<Pair<String, Double>>,
+        val buildDuration: Long,
+        val taskDurations: List<Pair<String, Long>>,
         val ext: BuildTimeTrackerPluginExtension
 )
